README updated

+++ src/pySymStat/quaternion.py
@@ -26,15 +26,17 @@ def quat_same(q1, q2, eps = 1e-3):
     return np.linalg.norm(q1 - q2) < eps or np.linalg.norm(q1 + q2) < eps
 
 def quat_rotate(q, v):
-    v = np.array([0, v[0], v[1], v[2]])
+    v = np.array([0, v[0], v[1], v[2]], dtype = np.float64)
     qv = quat_mult(quat_mult(q, v), quat_conj(q))
     return qv[1:]
 
 if __name__ == '__main__':
+
     q1 = np.random.randn(4)
     q1 /= np.linalg.norm(q1)
     q2 = np.random.randn(4)
     q2 /= np.linalg.norm(q2)
+
     v = np.random.randn(3)
     v /= np.linalg.norm(v)
